+++ local/airouter/classes/check/routerlisted.php
@@ -27,6 +27,11 @@ use core\check\result;
  */
 class routerlisted extends base {
     #[\Override]
+    protected function requires_instance(): bool {
+        return true;
+    }
+
+    #[\Override]
     protected function check_router(): result {
         if ($this->inspector->is_router_listed()) {
             return new result(result::OK, get_string('check:routerlisted:ok', 'local_airouter'));